pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgcaller"
version = "1.0.0"
description = "Modern, fast, and reliable Telegram group calls library with advanced features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TgCaller Team", email = "team@tgcaller.dev"}
]
maintainers = [
    {name = "TgCaller Team", email = "team@tgcaller.dev"}
]
keywords = [
    "telegram", "calls", "voip", "streaming", "audio", "video",
    "pytgcalls", "alternative", "modern", "fast", "reliable",
    "screen-sharing", "transcription", "youtube", "microphone",
    "bridged-calls", "filters", "advanced"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pyrogram>=2.0.106",
    "aiortc>=1.6.0",
    "aiofiles>=23.1.0",
    "aiohttp>=3.8.4",
    "requests>=2.31.0",
    "asyncio-throttle>=1.0.2",
    "loguru>=0.7.0",
    "psutil>=5.9.5",
    "pydantic>=1.10.8",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
media = [
    "ffmpeg-python>=0.2.0",
    "numpy>=1.24.0",
    "opencv-python>=4.7.0",
    "imageio>=2.28.0",
    "yt-dlp>=2023.6.22",
]
audio = [
    "pyaudio>=0.2.11",
    "soundfile>=0.12.1",
]
advanced = [
    "openai-whisper>=20231117",
    "mss>=9.0.1",
    "torch>=2.0.0",
]
all = [
    "tgcaller[dev,media,audio,advanced]"
]

[project.urls]
Homepage = "https://github.com/tgcaller/tgcaller"
Documentation = "https://tgcaller.readthedocs.io"
Repository = "https://github.com/tgcaller/tgcaller"
Issues = "https://github.com/tgcaller/tgcaller/issues"
Community = "https://t.me/tgcaller"

[project.scripts]
tgcaller = "tgcaller.cli:main"

[tool.setuptools.packages.find]
include = ["tgcaller*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"